Update README.md
Medcare is a website that helps to manage the related to health care. It manages the data related to all the deparments of healthcare. it contains the information about hospitals and take an appointment there online. With Medcare it is easy for searching best doctors for healthcare and medicine.
1) The figure shows the Home page of the website:
Medcare is a website that helps to manage the related to health care. It manages the data related to all the deparments of healthcare. it contains the information about hospitals and take an appointment there online. With Medcare it is easy for searching best doctors for healthcare and medicine.
1) The figure shows the Home page of the website: